Updates to make it work with EclipseLink and JPA-2.1 - Note it's compiled against a custom Oval-version (1.83-SNAPSHOT) to deal with getting field-values, see https://sourceforge.net/p/oval/discussion/488109/thread/e98d2f31/?limit=25#9214

+++ blog/src/main/scala/no/officenet/example/rpm/blog/domain/model/entities/BlogEntrySummary.scala
@@ -10,4 +10,4 @@ case class BlogEntrySummary(blogId: java.lang.Long,
 							title: String,
 							summary: String,
 							content: String,
-							numComments: Int)
+							numComments: Long)
